+++ app/models/slicers/numeric_slicer.rb
@@ -2,6 +2,6 @@
 
 module Slicers
   # Provides default methods for checking variables and saving to database.
-  class NumericSlicer < DefaultSlicer
+  class NumericSlicer < DomainSlicer
   end
 end