+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   - The Slice Expression Engine interprets the new Slice Context Free Grammar
     that allows users to build search templates to filter subjects and sheets
     based on one or more conditions
+- **Variable Changes**
+  - Numeric, integer, and calculated variables now link sheet values to domains
+    when there is a match
 
 ## 65.0.0 (September 10, 2018)